marimo.stop
- Add a function `stop(predicate, output)` that stops execution of a cell when
`predicate` is `False`

When `predicate` is `True`, this function stops execution of the
current cell and makes `output` its output. Any descendants of this cell
that were previously scheduled to run will not be run, and their defs will
be removed from program memory.

Example:

```python
mo.stop(form.value is None, mo.md("**Submit the form to continue.**"))
```

- Refactor runtime