Fix: Graceful exit when cell is executing (#253)
* Fix: Graceful exit when cell is executing

This PR fixes a bug in which the kernel process would not terminate
on SIGTERM if it was executing a cell. This left zombie marimo processes
on the user's system.

The fix is to no longer catch BaseException's in the cell runner, just
regular Exceptions and our special MarimoInterrupt and MarimoStopError
exceptions. This means that if a user types sys.exit(), or `raise
Keyboard Interrupt` in a cell, it will kill the kernel. The server isn't
aware that the kernel dies in these cases (which should be rare), but we
can fix that in a future change.

* remove test that no longer applies