Fix undefined behavior in kernel shutdown
Previously we were shutting the kernel down with sys.exit(). Exiting
with sys.exit() when the process was created after a fork may trigger
undefined behavior, since cleanup functions run twice, once for the
parent process and once for the child process.

This change uses os._exit() instead to kill the child process, as
recommended by Python docs.

This change also makes the SIGTERM handler re-entrant, to fix a bug
in which the virtual file registry shutdown method was being interrupted
upon receipt of a second SIGTERM. We terminate the process once
manually, and because the process is a daemon process it gets another
SIGTERM.