fix: [windows] refresh shouldn't kill server (#284)
This change fixes a bug in which refreshing in edit mode would kill the
server. Previously, a CTRL_BREAK event was sent to the kernel process
- but this killed the server, since the kernel process is in the same
process group as the parent. Instead, this change uses the os.terminate()
function to force the process to exit.

A better fix would figure out how to put the kernel process in its own
process group on Windows; that would let us send the ctrl-break and
still do cleanup in the kernel process. This is possible through the
subprocess module (creation_flags), but not through the multiprocessing
module as far as I can tell. To start a process with subprocess, though,
requires starting it as an external program, which is a extra (coding)
work. In a future change we can move toward that ...