fix: elminate bouncing of tooltip near end of editor
Prior to this fix, cursor tooltips would bounce distractingly
when typing near the end of a cell, as codemirror would think
the tooltip exceeded its space and therefore repositioned it.

This change eliminates that bounce by switching back to fixed
positioning for tooltips.

This change does _not_ regress selectable tooltip text (we had switched
from fixed to absolute to allow text to be selected, but this came with
downsides). To ensure text is still selectable, we make the Cell
responsible for clearing tooltips on blur, instead of the EditorView
(the Cell was already responsible for closing completions, for this
reason -- now it also clears tooltips).