fix: disallow `import *` in marimo notebooks
Python disallows `import *` everywhere except at the module-level. We
store marimo code in functions, so we can't support `import *`. Star
imports are also not great because they introduce globals that
marimo can't easily track

This fixes a bug in which a user could create a notebook with a `star`
import but then couldn't re-open it due to the syntax error.