fix: don't check server token in run mode
This change modifies the server to only validate the `ServerToken`
when running as a notebook (edit mode), and not when running as an app
(run mode).

The `ServerToken` was introduced to prevent a frontend started for one
notebook (eg, marimo edit locally) from accidentally clobbering/mutating
another notebook served on the same port but a different server. Since
run mode is read-only, no such risk exists.

This fix allows a frontend in run mode to connect to a server that died
and was restarted without refreshing the page.