fix: check frontend's app matches server's app in run mode
This change makes the server token in run mode a hash of the code,
to prevent frontends made for one app from trying to instantiate on
a server serving a different app. This still allows frontends
to connect to a server that died and came back as long as the server
is serving the same app.