Fix: kernel process should not be daemonic
The kernel process should not be daemonic, since daemon processes
can't create children (notebooks may use multiprocessing).

The process was previously made daemonic as a precaution to make sure it
terminated when the parent did. We have since hardened the code to make
sure that the kernel process is terminated, and shouldn't need the
daemon flag anymore.