Relax validation of args/returns on cells (#502)
* Relax validation of args/returns on cells

This change removes validation of a cell's args and returns on notebook
startup. The validation made spot-editing a notebook onerous, and
it's not really needed because marimo will auto-fix errors on save.

This also makes it easier to write a notebook from scratch in your
text editor, because you can just omit the refs and defs. Eg, you could
open the following file with marimo edit:

```python3
import marimo
app = marimo.App()

@app.cell
def imports():
  import marimo as mo

@app.cell
def greetings():
  mo.md("# Hello, world!")
```

and marimo will autofix it for you on save.

This change was motivated by feedback from users, who wanted to be able
to easily make edits to their notebooks in their editor of choice.

* Remove vestigial tests

* Revert "Remove vestigial tests"

This reverts commit 8992be5de385a7d4e57f8900e6e7cbf6dcdd9154.

* Tests