Improve README
Remove embedding video, based on feedback that it's hard
to understand -- too much going on.

Split out tutorial prose and GIFs from quickstart -- quickstart should
be as short as possible -- and add it as a separate section. This
separate section has simpler examples that very clearly explain what
marimo is.

The ordering is:

- Highlights
- Super quickstart [pip install marimo && marimo tutorial intro]
- Link to longer quickstart
- Overview of marimo (split out from quickstart)
- Quickstart