improvement: make footer more visible (#554)
This change makes the footer bar and the footer buttons more noticeable
by:

- adding a slight shadow
- increasing the size of the icons
- moving the help icon next to the panel icons (it's not a panel,
  but it's more visible there and the discongruity is ok)

Also tried experimenting with turning the icons black, or floating them
in the left as a sidebar instead of a footer. The sidebar could be a
better long-term solution -- right now we don't really have
any design elements to bring the user's eyes to the bottom of the
screen. But this combined with better docs is hopefully sufficient for
now ... can wait and see how users respond.