Implement standard in (backend)
This change implements support for reading from standard in on the
backend.

This is achieved by adding a new queue to each session/kernel pair, the
`input_queue`. The `input()` builtin and `sys.stdin.readline` function are
patched to send a control message to the frontend, requesting input.
They then pull an input from the input queue, blocking until an input is
available.

A new route is added, /api/kernel/stdin, that just accepts text and puts
it in the input queue.