fix: fix standard stream direction + multiprocessing
This change fixes a bad interaction between standard stream redirection
and multiprocessing, in which the forwarding thread would fail to join.

Instead of creating a new forwarding thread for each cell execution,
this change creates the forwarding threads at runtime start, and keeps
them alive for the duration of the kernel process. Similarly, instead
of creating a pipe for the redirection on each cell execution, this
change just creates the pipes once.