Take completion queue off the kernel's main thread ...
... and drain the completion queue on each completion.

This PR takes the completion queue off the kernel's main thread.
Instead, it adds a completion queue to the server's QueueManager, and
completion requests are enqueued by the server endpoint (so that they
reach the completion worker without going through the kernel).

As a result, completions now work even when the kernel is executing
code.

Additionally, this PR changes the completion worker to drain its queue
when retrieving completions, processing only the newest completion
requests and ignoring ones that it had received but has not yet
processed.

As a result, stale completions that take a long time to process (such as
for numpy, torch, tensorflow, ...) no longer waste the completion
worker's time (previously we would process all requests, even ones
the frontend no longer needed).