fix: remove multiprocessing mock for WASM
This change removes the multiprocessing mock in order to extend
compatibility to packages that import multiprocessing but don't use
offending features. In particular, with this fix it is now possible
to import `sklearn` in WASM notebooks.

Of the modules that were previously mocked, the only one that actually
fails to be imported under pyodide is `shared_memory`. Instead of
mocking this module, this PR updates our runtime code to just not import
this module when running under Pyodide.