improvement: hold on to unnamed UI elements
Changes the semantics of the UIRegistry to hold on to unnamed UI
elements.

This makes function call requests work on unnamed UI elements.

The registry now keeps track of all the UI elements registered to a
cell, and exposes a method to unregister them. The runtime unregisters
all UI elements for a cell when cell state is invalidated.