feat: run_button (#1514)
* feat: run_button

Adds `mo.ui.run_button()`, which can be used to trigger computations
when a button is pressed and ONLY when a button is pressed.

When the button is pressed, its value is set to `True` and its
descendants are run. After its descendants have run, marimo
automatically sets the button's value back to False. In this way,
cells can gate computations on `run_button.value` -- updates to other
variables won't spuriously cause the cell to run.

Caveat: The button's value is only reset to `False` when the runtime is
configured to autorun cells, since the lazy runtime doesn't know when
the button's descendants will/have been run. This should be fine in
practice, since `ui.run_button` has no real use in lazy kernels anyway.

* format

* improvement

* review comments