fix: uninstantiated logic (#1632)
This change fixes a bug in which cells were incorrectly marked as stale
on session startup/resume.

Previously, the frontend set a cell's `staleInputs` to `true` if
autorun-on-startup was disabled or if a cell didn't have an execution
time. But errored cells don't have execution times. And in general, setting
`staleInputs` is error-prone and only something the BE can determine.

This change instead initializes with `staleInputs=False`, letting
session replay override it as usual, and makes the check for whether
a cell is `uninstantiated` more robust.

Tested by manually playing around with a lot of edge cases.