improvement: less scroll shift
This change modifies scrolling to cells to use block nearest instead
of center, keeping the page as is if the element is already in view.

As a result, running shift enter, adding or deleting a cell
no longer scrolls the page (unless eg the created cell is not in the
viewport).

3 users have complained about our scroll being too jumpy, especially on
cell run -- this addresses that feedback.