perf: lazy imports when checking for datasets
This change adds an `imported()` method to the DependencyManager that
checks whether a module has been imported, and updates the
broadcast-datasets codepath to use this check instead of
DependencyManager.has().

Previously, running an arbitrary line of code (eg import marimo as mo)
could take 10+ seconds, as our table libraries (such as pandas) were all
being eagerly imported in order to check if a variable was of a table
type. Instead, this change only does the import and instance check if
the module has already been imported, which should be the case if an
instance of it is in memory.

With this change, import marimo as mo takes just ms.

The bottleneck was discovered with cProfile using the --profile-dir
development CLI arg:

  marimo edit --profile-dir profiles/ notebook.py