Improve: temporarily show code
This PR improves the UX of working with hidden cells. When code is
hidden, clicking on the eye icon will now temporarily show the code as
long as the editor is focused, and hide it again once the editor is
blurred.

In this way users can, for example, hide markdown code and later edit it
without permanently unhiding it.

To permanently unhide code, users can use the cell action menu or the
keyboard shortcut, as before.

This change is based on user feedback, and it also happens to match the
behavior in Observable notebooks.