Revert "improve: Move highlighted code to newly created cells (#2484)"
This reverts commit 25afc557aac442708dd416db6037eaa0e77bfb00.

After some experimentation we decided that moving highlighted code
to newly created cells is not intuitive. It's common to have code
highlighted without an intention to split the cell. Splitting is
destructive because it marks cells as stale, so it should be something
that is done explicitly by the user -- either by the split hotkey
or by simply copy pasting.