experimental: mo.ui.data_editor (#2567)
This adds an experimental ui element for editing a dataframe, using
`mo.ui.data_editor`.

* Add `mo.ui.data_editor` (BE plugin and FE plugin)
* This adds a dependency on `narwhals`. It's a lightweight library (no
deps itself) and does a lot of what we try to roll ourselves, by
creating a single dataframe interface. There is a lot of other code to
clean up by depending on this. This library will also be better tested
again more versions of polars/pandas/other-df-library
* We use `ag-grid` for the editing. It is missing key features that only
exist in the enterprise that we will likely want and force us to move to
something else. This includes: cell range selection, copy/paste ranges.
I don't expect this future migration to change the API, only
implementation details. I also don't expect the migration to be any more
expensive since it was quite easy to add ag-grid and is limited to one
file.


Current features:
* edit cells (text, long text, boolean, number)
* add rows

Later features to add:
* delete rows
* copy/paste
* range selection
* date cell