release marimo-base to test pypi (#2602)
This change adds a workflow that releases a marimo-base package to
TestPyPI.

marimo-base is marimo without static artifacts, which are not needed
when running on WASM.

Because Python build tools only allow one project per pyproject.toml, it
is not possible to do this by adding a build rule to hatch. Instead,
marimo-base is built by adapting the pyproject.toml in CI (this is what
CVXPY does).

With this change, the wheel size is reduced from `15M` to `614k`.