fix: consistent value for no column in config
Previously, the value for column in a single-column notebook was
sometimes null, sometimes undefined. This led to false positives
on notebookNeedsSave.

Without this fix, if you open a notebook and save it without making
any changes (experimental multi-column is disabled), the notebook
is marked as needing to be save for the remainder of the session, no
matter how many times you click "save".

I found that I also had to change arrayShallowEquals to lodash isEquals
for the equality check on config objects to work.