docs: reorganize guides (#2632)
Received feedback that our guide had become difficult to navigate, due
to a lack of logical progression, especially for students for whom many
of these concepts will be new. This PR reorganizes the guides to make
them easier to navigate and read in sequence.

Some guides were moved and links to them in the repo have been updated.
Once this is merged, will create redirects to fix links in other places
(such as blog posts, newsletters, ...).