fix: consistent value for no column in config (#2631)
Previously, the value for column in a single-column notebook was
sometimes null, sometimes undefined. This led to false positives on
notebookNeedsSave. This PR changes it so that when a column is not set,
the corresponding config value is `null` (which corresponds to `None` in
Python).

Without this fix, if you opened a notebook and saved it with or without
making any changes (experimental multi-column was disabled), the
notebook was marked as needing to be save for the remainder of the
session, no matter how many times you clicked "save".

I found that I also had to change arrayShallowEquals to lodash isEquals
for the equality check on config objects to work.