Add Panel support (#2719)
## 📝 Summary

This PR adds Panel support for Marimo.

## 🔍 Description of Changes

At a high-level this PR does a few things, including serializing Panel
components to JSON, setting up bi-directional communications and
registering formatters for HoloViz libraries including Panel components,
`param.rx` and HoloViews plots.

Specifically this PR consists of:

- Implements a TypeScript `PanelPlugin` which renders the JSON into the
DOM and sets up bi-directional comms
- A `panel` class that wraps Panel components, serializes the component
to JSON and includes Javascript for the extension(s) and sets up
bi-directional comms
- Formatters for HoloViz types

### Loading of Bokeh/Panel JS and extensions

Currently we load the Panel extension contents into a dynamically
created `<script>` tag which loads the required libraries. This works
well as far as I can tell and happens lazily only when the first Panel
component is displayed.

### Bi-directional Communication

The bi-directional communication happens via the existing `AnyWidget`
Comm wrapper. This isn't necessarily important and unfortunately we need
to wrap it in another layer of indirection which implements the
`pyviz_comms` APIs.

### Remaining Issues

- Displaying a component a second time seems to not work currently. Will
have to investigate why but simplest would be to simply disable the
caching
- We're importing `pyviz_comms` in the `from_panel` module, this
shouldn't happen.
- Based on my reading comms currently encode binary data as base64. I'd
love to see this improved (or to be corrected)
- Some refactoring in Panel so we don't have to duplicate so much
extension and serialization code here

## Example


![marimo_panel](https://github.com/user-attachments/assets/ba9f2efb-598c-49fd-8c8b-98571c3d8b33)

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://discord.gg/JE7nhX6mD8), or the community
[discussions](https://github.com/marimo-team/marimo/discussions)
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

@mscolnick as promised

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>