improvement: mpl proxy middleware (#2875)
Heavy lifting done by @maxtheman! 

This puts the mpl communication behind the main server URL and port, and
proxies request internally.

via @maxtheman:

## 📝 Summary

This represents the matplotlib fix to create a proxy middleware which
connects the main starlette process to the mpl starlette process.

Using this mapplotlib can be used in production on Docker for any kind
of interactive stuff.

See my demo app here, which is finally working:
https://causal-quartets.fly.dev/

I left the environmental variables in the codebase so that you could
customize the port or hosts for some reason in the future, but I removed
references to them from the documentation since I think that would be
more of a developer-facing thing anyway.

## 🔍 Description of Changes

1. Creates AsyncHTTPClient class to manage HTTP responses between the
two starlette processes without any dependencies (i.e., no httpx)
2. Fixes The WebSocket handling from the original implementation of the
proxy class. There were two things wrong: The websocket proxy was
dropping the figure url params required for matplotlib to find the
figure. There was also a bare Exception that was silently catching the
assertion failure in _mpl, which made diagnosis difficult.
3. adds 9 tests which successfully simulate and range the gamut across
all of the issues we were seeing and make sure that all of the HTTP and
WebSocket methods we're talking about work correctly as well as testing
it directly with the mpl middleware.

---------

Co-authored-by: Max Caldwell <2172753+maxtheman@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>