Allow numpy steps (#2958)
## 📝 Summary

In an attempt to improve the developer experience I figured it could be
nice to automatically cast Numpy arrays to lists inside of the slider
function. Otherwise the user would need to cast to a list themselves in
order not to see an error.

Per discussion on
[Discord](https://discord.com/channels/1059888774789730424/1310581204520665200/1310601852211036272)

## 🔍 Description of Changes

I tried loading numpy lazily inside of the `mo.ui.slider` call. If the
`steps` are not a list, we can proceed to check if it is a numpy array.
If it is, make the change, leave as-is otherwise to raise appropriate
errors as one would normally.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

@akshayka OR @mscolnick

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>