feat: override configuration with pyproject.toml (#3071)
Fixes #3069 
Related to #2922, but maybe doesn't close it quite yet. 

This adds `pyproject.toml` configuration overrides, so a collection of
notebook can pick up the same marimo configuration. This overrides the
user's configuration. You can only edit this in the pyproject.toml - and
not in the UI. When there is an override, we show in the UI the fields
are overridden.

----

### Overriding settings with pyproject.toml

You can override user configuration settings with a `pyproject.toml`
file. This
is useful for sharing configurations across teams or ensuring
consistency across
notebooks. You must edit the `pyproject.toml` file directly to override
settings.

For example, the following `pyproject.toml` file overrides the
`autosave` setting
in the user configuration:

```toml
[tool.marimo.save]
autosave = "after_delay"
```

You can override any user configuration setting in this way. To find
these settings run `marimo config show`.

![Screenshot 2024-12-06 at 1 41
54 PM](https://github.com/user-attachments/assets/e711f196-d553-47e6-860b-66cea98fe3d7)
![Screenshot 2024-12-06 at 1 41
40 PM](https://github.com/user-attachments/assets/a6cce068-340f-403f-a8c8-13fcbef4fefd)

![Screenshot 2024-12-06 at 1 42
31 PM](https://github.com/user-attachments/assets/bdb4cdc6-e07c-40e1-9d09-fcc08d8b7199)