fix: .schema perf improvements (#3108)
Fixes #3107 

`.schema` from narwhals or polars is slow if not cached. I didn't expect
these to be slow. Due to a refactor we called this from the inner loop
each time. This PR fixes that and adds a test.

Will do some more perf testing on various methods, that get called a
lot.