feat: support for Azure OpenAI with AI Assistant (#3120)
## 📝 Summary

Resolves marimo-team/marimo#3117

## 🔍 Description of Changes

As discussed in marimo-team/marimo#3117, I'm inferring whether to
instantiate an AzureOpenAI client or a regular OpenAI client based on
the format of the base url. I made this change consistently in two
places: once for the AI assistant and one for the `llm` class in Chat.

Note one other change: when streaming, the first response for the
"choices" array can sometimes be empty (not sure if this is an Azure
thing, or in general). So this also includes a one-line change to handle
that case.

I have verified locally that this change allows the AI assistant to
work, as well as the [openai
example](https://github.com/marimo-team/marimo/blob/main/examples/ai/chat/openai_example.py)
assuming you pass in the same base url.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@akshayka OR @mscolnick