fix: gracefully close dead sessions in edit mode (#3203)
Fixes #3187.

Previously we brought down the whole edit server if an edit-mode kernel
died unexpectedly. This used to make sense when marimo was 1 server, 1
notebook, but it no longer makes sense today, as it causes the user to
lose all their work in other notebooks opened by the server too.

Instead, this change makes it so that the session is closed (and removed
from the manager), but the server is still kept alive.