Nitpick: prefer --directory flag over cd (#3215)
## 📝 Summary

Right now when you export to html-wasm this line is shown: 

```
To run the exported notebook, use:
  cd docs/index.html && python -m http.server
```

It works, but the `cd` is a side effect. Might be nice to be able to do
without it?

## 🔍 Description of Changes

You can use a `--directory` flag for `python -m http.server` instead.
Might be preferable? This way you don't have to worry about cd'ing in
and out.

## 📜 Reviewers


@akshayka OR @mscolnick