examples: ui element usage examples (#3254)
Adds a collection of small self-contained examples for our various UI
elements. In a follow-up, it might make sense to have our API docs point
to these examples as a the source of truth (instead of inlining
`marimo-embed`).

Adds a pointer to these examples and the API docs to `marimo tutorial
intro`.