feat: apply pytest rewrites for better testing (#3294)
merry merry 🎄 

---

Slight improvement on pytest errors, by manually running pytest's
assertion rewrite magic. So that takes us from this:


![image](https://github.com/user-attachments/assets/e2ddf88b-a6c1-4f89-a98b-20f70b0e8c6d)

to this


![image](https://github.com/user-attachments/assets/317a9d4b-faa1-4d9b-9f54-fa5e8c747e39)

Also a little fun- this rewrite can be applied in general and shown in
editor.


![image](https://github.com/user-attachments/assets/18934182-732a-4dad-8e91-12b9abd32f7c)

Not sure the overhead this might have- so not sure about setting it
generally, but could set `test_rewrite=True` if the cell name starts
with `test_`. Wanted to run this by you all first


@akshayka OR @mscolnick