fix: partially instantiate notebooks when autorun on startup is disabled (#3316)
When autorun on startup is disabled, running a cell should only run its
uninstantiated ancestors, and in particular it shouldn't run the
uninstantiated descendants of of uninstantiated ancestors.

This lets the notebook author incrementally run parts of their notebook
on startup, even those that include UI elements, without making the
entire notebook lazy.

Previously, running a cell would collect uninstantiated ancestors but
also uninstantiated descendants, which would often be the entire
notebook since uninstantiated ancestors often include imports.

This fix was implemented by keeping track of which execution requests
are uninstantiated. This data structure is pruned as cells are run or
deleted.