feat: add auto-fix for multiple definctions and popular imports (#3343)
Part of #3110

This adds two autofixes for different marimo errors. This opens the door
for more autofixes as well. Right now this happens on all the frontend
because 1) the code transformations are not that difficult and 2) still
requires frontend logic like focusing the editor or adding a new cell.
We could move some of this to the backend but would prefer to put the
auto-fixes into a a proper linting framework + LSP.

The autofixes:
* multiple definitions -> Make this cell's variables local by wrapping
the cell in a function.
* missing name -> Add a new cell for the missing import


![image](https://github.com/user-attachments/assets/2d1d7c15-8917-4cd0-b359-8ef27f779aef)