feat: add /public to serve public files in a directory next to the notebook (#3351)
This lets users grab images or data from `/public`. Any data inside the
/public folder next to the notebook will accessible via the `/public`
route

This pull request introduces functionality to serve files from a public
directory within a notebook's directory and adds security measures to
ensure safe file access. It also includes tests to verify the new
functionality.

We need this since we serve multiple notebooks from the same server, so
we need to find the location of the notebook. But since these are GET
request, we cannot use headers. And since we don't control the URLs
(comes from the user), we cannot add query-params.

So we can use service workers or cookies. This implementation chooses
service workers.

* Added a new endpoint to serve files from a `/public` directory within
a notebook's directory, including a service worker to manage these
requests.
* Introduced a service worker that registers with the notebook ID and
adds it to request headers for files in the `/public` directory.