fix: Add missing open bracket in docs/faq.md (#3446)
## 📝 Summary

This PR adds a missing open bracket in the FAQ that broke one of the
links.

## 🔍 Description of Changes

I added the open bracket

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected - I
have manually verified that the link links to a site that actually
exists.

## 📜 Reviewers

@akshayka OR @mscolnick