Support shells on windows (with extensions)
Even though the value of $SHELL is just "bash" at the command line, it
is the full path to the executable in the program. (i.e.
'c:\foo\bar\bash.exe'). The "name" attribute returns "bash.exe" in that
case, which doesn't match "bash". Fortunately, pathlib.Path also offers
a "stem" attribute which strips the extension. \o/