fix: conversion of ipynb files with multiple definitions (#3458)
The `NameTransformer` wasn't visiting node descendants, leading to
incomplete transformations.

Before, when remapping `x` to a local:

```python
x = 1
y = foo(x + 1)
```

was yielding

```python
x = 1
y = foo(x + 1)
```

With this change, it now yields

```python
_x = 1
y = foo(_x + 1)
```

This change also modifies an old test that was testing that the
`NameTransformer` respected scoping logic. But it never actually did, or
rather only did so because of a bug in which it wasn't visiting entire
subtrees. Nothing relies on the `NameTransformer` respecting scoping in
our codebase.