fix: fallback to oringinal string when cannot parse formatted number (#3529)
Fixes #3528

When we add formatting to a column, we still say the data-type is a
float/int, but we actually pass a string. So we need to fallback
gracefully if so.

It's likely a better fix is letting the frontend know that the datatype
has changed if it was formatted to a string.