feat: add mo.ui.app_meta().request (#3619)
This adds `mo.ui.app_meta().request` object. This object is updated from
the last request that ran the cell (`instantiate`, `run`, or
`set_ui_element`).

This passes headers, queryparams, and the user object (if there is one).
We omit `state` and `session` since that may not want to be exposed