fix: kernel globals memory leak in run mode
This change fixes a memory leak in run mode in which a kernel's globals
memory appeared to not be freed on session (thread) exit. However,
another (smaller) leak appears to still exist.

This change also makes sure to stop the stream's buffered writer thread when
kernel exits.

Also, add a TODO for performance of VariableValue construction --
calling str() on some values, such as a byte array of size 1GB, can make
memory usage spike to 25GB.