fix: date range not defaulting to `start` and `stop` (#3651)
## 📝 Summary

`date_range` should default to `start` and `stop` according to the docs
but it doesn't. This fixes this.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.


## Pixi

Even though you did a nice job with encapsulating work with `hatch` and
`make` I was still surprised how many steps are needed to get a working
setup.
With pixi you could install all prerequisites, and replace both hatch
and make. Let me know if you'd be interested in a PR for that.

@akshayka OR @mscolnick