improvement: 'open in editor' via file-explorer (#3766)
This is somewhat useful now, but will be more useful with the LSP
integration (#3414). We can use this so that cmd+click will go to the
numpy file directly in your editor of choice.

TODO:
- [x] ~~this works well when `$EDITOR` is `code` or `cursor` since focus
is sent to that application. But not when `$EDITOR` is `vim`. This also
works well when there is no editor configured.~~ For now, just skipping
over terminal editors and calling `open`