feat: Session Schema and Serialize outputs  (#3807)
Fixes #3082

This adds support for serializing/deserializing part of the SessionView
so that we can re-hyrdate the outputs on startup.
While this looks like it may conflict with caching (`mo.cache`), it has
a much different use-case. This is a much simpler version for different
tradeoffs:
- only storing output (not variables or previous runs)
- everything is stale on startup

This acts like Jupyter's output storage, but storage in a different
file.

This adds:
- a code-first schema `NotebookSessionV1`
- we generate the openai spec (for easier code-reviews) and a GitHub
Action to detect breaking changes
- serialization/deserialization
- `AsyncBackgroundTask` util that our `SessionCacheWriter` runs on
- `Session` starts this task, only when auto-instantiate if off

Note:
- timings are not handled in the ser/de or tests (yet), will handle in a
followup
- If the path is `foo/bar/baz.py`, the cache file is →
`foo/bar/__marimo__/session/baz.py.json`

In a followup:
- Handle timings
- WASM support