compat: Register webbrowser.open hook for pyodide (#3864)
## 📝 Summary

`import webbrowser.open` seems to be broken in WASM (which is fine, I
think opening pages is potentially dangerous)- so this loads in a hook
to insert an iframe into output if webbrowser is called instead of
failing. As a result, the following code now works

```python
import antigravity
```

but with some marimo tweaks that might make for good social engagement
on 25/04/01


![image](https://github.com/user-attachments/assets/8721a668-516b-40c8-9ae1-86056d58151a)

(PS: this needs a static upload)

Had a spare hour and noticed it when looking through cpython/Lib
Let me know if it's too irreverent and I can walk back the feature- but
seems enabling this seems like full python support to me

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@akshayka OR @mscolnick

---------

Co-authored-by: Myles Scolnick <myles@marimo.io>